# Feature/optimize draw oled buffer