Thank you This Great Program !

I modified and Improved these .
So I pull request !

* Add I2C bus define [d760234](https://github.com/FREEWING-JP/ESP32_BadApple_MP3/commit/d76023401bc675fd5c6eb35f52aac7265b5c94a3)
* Add OLED setBrightness [63d5239](https://github.com/FREEWING-JP/ESP32_BadApple_MP3/commit/63d5239a761fd879e50b2694b8535b4e9c523e25)
* More Accurate frame rate [dedd45f](https://github.com/FREEWING-JP/ESP32_BadApple_MP3/commit/dedd45f6c43eaf9b8bf4342e965009c5f972e17b)
* Optimize Logic []()
* I2C SCLK 700kHz to 4MHz [fab9458](https://github.com/FREEWING-JP/ESP32_BadApple_MP3/commit/fab9458897799459870bfababf3b3ce429da2178)
* Optimize Direct Draw OLED buffer [dbcd6fe](https://github.com/FREEWING-JP/ESP32_BadApple_MP3/commit/dbcd6fe269458c6afc99214acd019e28443d378c)
* etc. See each Commit Message

### Playing Time Comparison
Movie 6573 frame / 30fps = 219.1 sec  

Comment|commit hash | Time
-- | -- | --
Original|a6f330c|3:36.941
mod Add Frame counter|1974d05|3:39:094
mod Optimize Direct Draw OLED buffer|395e4d9|3:39.090

### No Wait Speed Comparison
Comment|commit hash | Time |Speed Up
-- | -- | -- | --
Original|a6f330c|1:37.886|--
mod Add Frame counter|1974d05|1:39.652|-2 sec
mod Optimize int16_t to int32_t|c33224c|1:38.390|-1 sec
mod I2C SCLK to 4MHz|2ee30c0|1:21.606|16 sec
mod More Optimize Direct Draw OLED buffer|994f92b|1:03.274|34 sec
mod Del unnecessary Initialize|158a3a2|1:02.283|35 sec